requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
plotly>=5.0.0
python-dateutil>=2.8.0
click>=8.0.0
flask>=3.0.0
flask-cors>=4.0.0
